[REFACTOR] Update react structure about comment
