+++ client/src/components/PostView/Comment/CommentContainer/CommentContainer.jsx
@@ -1,9 +1,8 @@
 import React, { useEffect, useCallback, useState } from 'react';
-import styled from 'styled-components';
 import { useSelector } from 'react-redux';
-import { getCommentListRead } from '../../../network/comment/http';
-import { getCookie } from '../../../utils/cookie';
-import CommentWrapper from './CommentWrapper';
+import { getCommentListRead } from '../../../../network/comment/http';
+import { getCookie } from '../../../../utils/cookie';
+import CommentContainerPresenter from './CommentContainer.style';
 
 export default function CommentContainer({ postId }) {
   const tkn = getCookie('users');
@@ -31,21 +30,9 @@ export default function CommentContainer({ postId }) {
   const uploadingComments = createdComments.filter((el) => el.parent_id == null);
 
   return (
-    <CommentList>
-      {uploadingComments.length !== 0 &&
-        uploadingComments
-          .sort((a, b) => {
-            return a.createAt > b.createAt ? -1 : a.create < b.create ? 1 : 0;
-          })
-          .map((el) => <CommentWrapper key={el._id} cmntData={el} />)}
-      {loadedComments &&
-        loadedComments
-          .sort((a, b) => {
-            return a.createAt > b.createAt ? -1 : a.create < b.create ? 1 : 0;
-          })
-          .map((el) => <CommentWrapper key={el._id} cmntData={el} />)}
-    </CommentList>
+    <CommentContainerPresenter
+      uploadingComments={uploadingComments}
+      loadedComments={loadedComments}
+    />
   );
 }
-
-const CommentList = styled.div``;
