+++ client/src/components/PostView/Comment/PostComment/PostComment.jsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import PostCommentPresenter from './PostComment.style';
+
+export default function PostComment({ postId }) {
+  return <PostCommentPresenter postId={postId} />;
+}
